pyproject.toml
[project]
name = "structured-output-benchmark"
version = "0.1.0"
description = "Benchmark various structured output frameworks with identical inputs"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.131.0",
    "guardrails-ai>=0.9.0",
    "instructor>=1.14.5",
    "langchain-openai>=1.1.10",
    "loguru>=0.7.3",
    "marvin>=3.2.6",
    "mirascope[openai]>=2.2.2",
    "openai>=2.21.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0.3",
    "uvicorn[standard]>=0.41.0",
]
